Break Away Overlay from ModalBox
