+++ src/index.ts
@@ -74,6 +74,7 @@ export { default as Loader } from "./components/Loader";
 export { default as LoginWrapper } from "./components/LoginWrapper";
 export { default as MDSCacheProvider } from "./components/MDSCacheProvider";
 export { default as ModalBox } from "./components/ModalBox";
+export { default as Overlay } from "./components/Overlay";
 export { default as NotificationAlert } from "./components/NotificationAlert";
 export { default as NotificationCount } from "./components/NotificationCount";
 export { default as Notifications } from "./components/Notifications";
@@ -132,6 +133,7 @@ export * from "./components/InputLabel/InputLabel.types";
 export * from "./components/Link/Link.types";
 export * from "./components/LinkButton/LinkButton.types";
 export * from "./components/ModalBox/ModalBox.types";
+export * from "./components/Overlay/Overlay.types";
 export * from "./components/NotificationAlert/NotificationAlert.types";
 export * from "./components/Notifications/Notifications.types";
 export * from "./components/PageHeader/PageHeader.types";
